+++ frontend/src/mocks/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw/browser';
 
-import { handlers } from './handlers';
+import { handlers } from '@/mocks/handlers';
 
 export const worker = setupWorker(...handlers);
